requirements.txt
Flask==3.0.0
google-generativeai==0.3.0
pdf2image==1.16.3
python-docx==0.8.11
reportlab==4.1.0
python-dotenv==1.0.0